refactor(BaseDomain): BaseDomain사용
