+++ src/main/java/nextstep/courses/domain/Course.java
@@ -1,20 +1,15 @@
 package nextstep.courses.domain;
 
+import nextstep.common.domain.BaseDomain;
 import nextstep.session.domain.Session;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class Course {
-    private Long id;
-
+public class Course extends BaseDomain {
     private String title;
 
     private Long creatorId;
-
-    private LocalDateTime createdAt;
-
-    private LocalDateTime updatedAt;
 
     private List<Session> sessions;
 
@@ -26,11 +21,9 @@ public class Course {
     }
 
     public Course(Long id, String title, Long creatorId, LocalDateTime createdAt, LocalDateTime updatedAt) {
-        this.id = id;
+        super(id, createdAt, updatedAt);
         this.title = title;
         this.creatorId = creatorId;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
     }
 
     public String getTitle() {
@@ -41,18 +34,12 @@ public class Course {
         return creatorId;
     }
 
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
     @Override
     public String toString() {
         return "Course{" +
-                "id=" + id +
-                ", title='" + title + '\'' +
+                "title='" + title + '\'' +
                 ", creatorId=" + creatorId +
-                ", createdAt=" + createdAt +
-                ", updatedAt=" + updatedAt +
+                ", sessions=" + sessions +
                 '}';
     }
