+++ src/main/java/nextstep/session/domain/Session.java
@@ -1,21 +1,16 @@
 package nextstep.session.domain;
 
+import nextstep.common.domain.BaseDomain;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public abstract class Session {
+public abstract class Session extends BaseDomain {
     private static final SessionStatus DEFAULT_SESSION_STATUS = SessionStatus.PREPARING;
 
-    private Long id;
-
     private Long creatorId;
-
-    private LocalDateTime createdAt = LocalDateTime.now();
-
-    private LocalDateTime updatedAt = LocalDateTime.now();
 
     private LocalDate startDate;
 
@@ -23,14 +18,20 @@ public abstract class Session {
 
     private SessionImage sessionImage;
 
-    protected SessionStatus sessionStatus = DEFAULT_SESSION_STATUS;
+    protected SessionStatus sessionStatus;
     protected SessionStudents students = new SessionStudents();
 
     public Session(Long creatorId, LocalDate startDate, LocalDate endDate, SessionImage sessionImage) {
+        this(0L, LocalDateTime.now(), null, creatorId, startDate, endDate, sessionImage, DEFAULT_SESSION_STATUS);
+    }
+
+    public Session(Long id, LocalDateTime createdAt, LocalDateTime updatedAt, Long creatorId, LocalDate startDate, LocalDate endDate, SessionImage sessionImage, SessionStatus sessionStatus) {
+        super(id, createdAt, updatedAt);
         this.creatorId = creatorId;
         this.startDate = startDate;
         this.endDate = endDate;
         this.sessionImage = sessionImage;
+        this.sessionStatus = sessionStatus;
     }
 
     public void enroll(NsUser user) {
